requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
requests
passlib[bcrypt]
python-multipart
